split the process model out of Async\ into Process\, and let plain pcntl_signal work
Layering fix, prompted by the right question: is a signal really part of the
async namespace? Mostly it is not.

- Process\ (new, in the pcntl tier) takes fork/pid/ppid/workers/supervise. None
  of that runs a scheduler -- supervise() forks, reaps, restarts and forwards
  SIGTERM, and Async\supervise() implied a runtime that supervision never has.
  The process model sits UNDER concurrency, not inside it.
- Async\onSignal is gone. The scheduler notices a non-empty pcntl registry and
  starts the pump itself, so ordinary `pcntl_signal(SIGHUP, ...)` just works
  inside async() -- one way to register a handler, and it is the standard one.
- Async\pid() is gone (it was posix_getpid() under another name), along with
  the duplicate sys_fork/sys_getpid FFI declarations in async.php.
- Async\shutdownOn() stays, and is now the only signal-aware thing in Async\.
  It earns that by being about SCOPES: it cancels the root scope, and the rest
  falls out of cancellation already being structured.

Two things this shook out:

- The lazy pump start has to be checked immediately BEFORE the blocking reactor
  wait, not at the top of the loop. A worker parked in accept() reaches the top
  again only once the reactor returns, so an idle one started no pump and never
  saw its SIGTERM -- in the four-worker demo exactly one worker drained (the one
  curl had woken) and three hung. Checked before the block, the pump lands in
  the ready queue and its own delay() then bounds the wait.
- The gate cannot use definedFunctions() for this tier any more: it now defines
  fork/pid/workers/supervise, which are names a program may well own. pcntl_*
  and posix_* are prefixed and stay gated on the call; Process\ is gated on the
  qualifier, exactly as Async\ is. Verified: a program defining its own
  workers() and supervise() compiles and runs untouched.

poc/async/server.php, http_server.php, http_transparent.php and async_signal
move to Process\; async_signal now registers with plain pcntl_signal to prove
the auto-pump. async 11/11, pcntl 1/1 (difftest MATCH), fiber 9/9; the four
supervised workers all drain and the supervisor exits.
